requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.8
dataclasses-json==0.5.6
idna==3.3
isodate==0.6.0
marshmallow==3.14.1
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
oauthlib==3.1.1
python-decouple==3.5
python-youtube==0.8.1
pytube==11.0.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
typing-extensions==4.0.0
typing-inspect==0.7.1
urllib3==1.26.7
